requirements.txt
sqlalchemy
openai
pydantic_settings
pymilvus
frontend
pymupdf
python-pptx
openpyxl
python-docx
unidecode
igraph
leidenalg
pyvis
langchain_community
crawl4ai
transformers
SentencePiece
timm
scikit-learn
opencv-python
elasticsearch==8.17.2
colpali-engine
SpeechRecognition
streamlit
pydub
numpy 
elasticsearch==8.17.2
plotly==6.0.1
nbformat >= 4.2.0
kaleido==0.2.1
streamlit==1.44.0
mistralai
ecologits
rapidfuzz
duckduckgo-search
